@@ -1,13 +1,14 @@
-import 'package:ez_localization/ez_localization.dart';
+import 'dart:io';
+
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:unicaen_timetable/model/lessons/authentication/state.dart';
-import 'package:unicaen_timetable/model/lessons/repository.dart';
-import 'package:unicaen_timetable/model/lessons/user/repository.dart';
-import 'package:unicaen_timetable/model/lessons/user/test.dart';
-import 'package:unicaen_timetable/model/lessons/user/user.dart';
-import 'package:unicaen_timetable/model/settings/settings.dart';
-import 'package:unicaen_timetable/utils/progress_dialog.dart';
+import 'package:unicaen_timetable/i18n/translations.g.dart';
+import 'package:unicaen_timetable/model/settings/calendar.dart';
+import 'package:unicaen_timetable/model/user/calendar.dart';
+import 'package:unicaen_timetable/model/user/user.dart';
+import 'package:unicaen_timetable/utils/lesson_download.dart';
+import 'package:unicaen_timetable/widgets/centered_circular_progress_indicator.dart';
+import 'package:unicaen_timetable/widgets/dialogs/input.dart';
 
 /// The user login dialog.
 class LoginDialog extends ConsumerStatefulWidget {
@@ -27,6 +28,7 @@ class LoginDialog extends ConsumerStatefulWidget {
   static Future<bool> show(BuildContext context, {bool synchronizeAfterLogin = false}) async {
     bool? result = await showDialog<bool>(
       context: context,
+      barrierDismissible: false,
       builder: (context) => LoginDialog(synchronizeAfterLogin: synchronizeAfterLogin),
     );
     return result ?? false;
@@ -38,139 +40,215 @@ class _LoginDialogState extends ConsumerState<LoginDialog> {
   /// The current form key.
   GlobalKey<FormState> formKey = GlobalKey<FormState>();
 
-  /// The current focus node.
-  FocusNode focusNode = FocusNode();
-
   /// The username text controller.
-  late TextEditingController usernameController;
+  late TextEditingController usernameController = TextEditingController();
 
   /// The password text controller.
-  late TextEditingController passwordController;
+  late TextEditingController passwordController = TextEditingController();
 
-  /// The current login result state.
-  RequestResultState? loginResultState;
+  /// The server text controller.
+  late TextEditingController serverAddressController = TextEditingController();
+
+  /// The calendar name text controller.
+  late TextEditingController calendarNameController = TextEditingController();
+
+  /// The additional parameters text controller.
+  late TextEditingController additionalParametersController = TextEditingController();
+
+  /// Whether a login has been requested.
+  bool waiting = false;
+
+  /// Whether the login button is enabled.
+  bool canLogin = false;
+
+  /// Whether to display more settings.
+  bool moreSettings = false;
+
+  /// The current login result.
+  int? loginHttpResponseCode;
 
   @override
   void initState() {
     super.initState();
 
-    usernameController = TextEditingController();
-    passwordController = TextEditingController();
-
-    WidgetsBinding.instance.addPostFrameCallback((_) {
-      User? user = ref.read(userRepositoryProvider).user;
+    WidgetsBinding.instance.addPostFrameCallback((_) async {
+      User? user = await ref.read(userProvider.future);
       if (user != null) {
         setState(() {
           usernameController.text = user.username;
           passwordController.text = user.password;
+          canLogin = formKey.currentState?.validate() == true;
         });
       }
     });
   }
 
   @override
-  Widget build(BuildContext context) {
-    List<Widget> children = [
-      Text(context.getString('dialogs.login.username')),
-      TextFormField(
-        decoration: InputDecoration(hintText: context.getString('dialogs.login.username_hint')),
-        autocorrect: false,
-        validator: (value) => value == null || value.isEmpty ? context.getString('other.field_empty') : null,
-        controller: usernameController,
-        textInputAction: TextInputAction.next,
-        onFieldSubmitted: (_) => FocusScope.of(context).requestFocus(focusNode),
-      ),
-      Padding(
-        padding: const EdgeInsets.only(top: 20),
-        child: Text(context.getString('dialogs.login.password')),
-      ),
-      TextFormField(
-        decoration: InputDecoration(hintText: context.getString('dialogs.login.password_hint')),
-        autocorrect: false,
-        obscureText: true,
-        keyboardType: TextInputType.visiblePassword,
-        validator: (value) => value == null || value.isEmpty ? context.getString('other.field_empty') : null,
-        controller: passwordController,
-        textInputAction: TextInputAction.done,
-        focusNode: focusNode,
-      ),
-    ];
-
-    if (loginResultState != null) {
-      children.add(createErrorMessage());
-    }
-
-    return AlertDialog(
-      title: Text(context.getString('dialogs.login.title')),
-      content: SingleChildScrollView(
-        child: Form(
-          key: formKey,
-          child: Column(
-            mainAxisSize: MainAxisSize.min,
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: children,
-          ),
-        ),
-      ),
-      actions: [
-        TextButton(
-          onPressed: () => Navigator.pop(context, false),
-          child: Text(MaterialLocalizations.of(context).cancelButtonLabel.toUpperCase()),
-        ),
-        TextButton(
-          onPressed: () => onLoginButtonPressed(context),
-          child: Text(context.getString('dialogs.login.login').toUpperCase()),
-        ),
-      ],
-    );
-  }
-
-  /// Creates the error message widget.
-  Widget createErrorMessage() => Padding(
-        padding: const EdgeInsets.only(top: 20),
-        child: Text(
-          context.getString('dialogs.login.errors.${loginResultState!.id}'),
-          style: TextStyle(color: Colors.red[700]),
-        ),
+  Widget build(BuildContext context) => AlertDialog(
+        title: Text(translations.dialogs.login.title),
+        content: waiting
+            ? const CenteredCircularProgressIndicator()
+            : Form(
+                key: formKey,
+                child: Column(
+                  mainAxisSize: MainAxisSize.min,
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    Text(translations.dialogs.login.username),
+                    TextFormField(
+                      decoration: InputDecoration(hintText: translations.dialogs.login.usernameHint),
+                      autocorrect: false,
+                      onChanged: refreshLogin,
+                      validator: TextInputDialog.validateNotEmpty,
+                      controller: usernameController,
+                      textInputAction: TextInputAction.next,
+                      autovalidateMode: AutovalidateMode.onUserInteraction,
+                      autofocus: true,
+                    ),
+                    Padding(
+                      padding: const EdgeInsets.only(top: 20),
+                      child: Text(translations.dialogs.login.password),
+                    ),
+                    TextFormField(
+                      decoration: InputDecoration(hintText: translations.dialogs.login.passwordHint),
+                      autocorrect: false,
+                      obscureText: true,
+                      keyboardType: TextInputType.visiblePassword,
+                      onChanged: refreshLogin,
+                      validator: TextInputDialog.validateNotEmpty,
+                      controller: passwordController,
+                      autovalidateMode: AutovalidateMode.onUserInteraction,
+                      textInputAction: moreSettings ? TextInputAction.next : TextInputAction.done,
+                    ),
+                    if (loginHttpResponseCode != null)
+                      Padding(
+                        padding: const EdgeInsets.only(top: 20),
+                        child: Text(
+                          switch (loginHttpResponseCode) {
+                            HttpStatus.notFound => translations.dialogs.login.errors.notFound,
+                            HttpStatus.unauthorized => translations.dialogs.login.errors.unauthorized,
+                            _ => translations.dialogs.login.errors.genericError,
+                          },
+                          style: TextStyle(color: Colors.red.shade700),
+                        ),
+                      ),
+                    if (moreSettings)
+                      ...[
+                        Padding(
+                          padding: const EdgeInsets.only(top: 20),
+                          child: Text(translations.dialogs.login.moreSettings.server),
+                        ),
+                        TextFormField(
+                          decoration: const InputDecoration(hintText: kDefaultServer),
+                          autocorrect: false,
+                          keyboardType: TextInputType.url,
+                          onChanged: refreshLogin,
+                          validator: TextInputDialog.validateNotEmpty,
+                          controller: serverAddressController,
+                          autovalidateMode: AutovalidateMode.onUserInteraction,
+                          textInputAction: TextInputAction.next,
+                        ),
+                        Padding(
+                          padding: const EdgeInsets.only(top: 20),
+                          child: Text(translations.dialogs.login.moreSettings.calendarName),
+                        ),
+                        TextFormField(
+                          decoration: const InputDecoration(hintText: kDefaultCalendarName),
+                          autocorrect: false,
+                          keyboardType: TextInputType.text,
+                          onChanged: refreshLogin,
+                          validator: TextInputDialog.validateNotEmpty,
+                          controller: calendarNameController,
+                          autovalidateMode: AutovalidateMode.onUserInteraction,
+                          textInputAction: TextInputAction.next,
+                        ),
+                        Padding(
+                          padding: const EdgeInsets.only(top: 20),
+                          child: Text(translations.dialogs.login.moreSettings.additionalParameters),
+                        ),
+                        TextFormField(
+                          decoration: const InputDecoration(hintText: kDefaultAdditionalParameters),
+                          autocorrect: false,
+                          onChanged: refreshLogin,
+                          keyboardType: TextInputType.url,
+                          controller: additionalParametersController,
+                          autovalidateMode: AutovalidateMode.onUserInteraction,
+                          textInputAction: TextInputAction.done,
+                        ),
+                      ],
+                  ],
+                ),
+              ),
+        scrollable: true,
+        actions: waiting
+            ? []
+            : [
+                if (loginHttpResponseCode != null && !moreSettings)
+                  TextButton(
+                    onPressed: () {
+                      setState(() => moreSettings = true);
+                    },
+                    child: Text(translations.dialogs.login.moreSettings.button),
+                  ),
+                TextButton(
+                  onPressed: () => Navigator.pop(context, false),
+                  child: Text(MaterialLocalizations.of(context).cancelButtonLabel),
+                ),
+                TextButton(
+                  onPressed: canLogin ? (onLoginButtonPressed) : null,
+                  child: Text(translations.dialogs.login.login),
+                ),
+              ],
       );
 
+  /// Refreshes whether the login button is enabled.
+  void refreshLogin(String? _) => setState(() => canLogin = formKey.currentState?.validate() == true);
+
+  @override
+  void dispose() {
+    usernameController.dispose();
+    passwordController.dispose();
+    serverAddressController.dispose();
+    calendarNameController.dispose();
+    additionalParametersController.dispose();
+    super.dispose();
+  }
+
   /// Triggered when the login button has been pressed.
-  void onLoginButtonPressed(BuildContext context) async {
-    if (!(formKey.currentState?.validate() ?? false)) {
+  void onLoginButtonPressed() async {
+    if (!canLogin) {
       return;
     }
 
-    ProgressDialog.show(context);
+    setState(() => waiting = true);
 
-    SettingsModel settingsModel = ref.read(settingsModelProvider);
-    UserRepository userRepository = ref.read(userRepositoryProvider);
+    if (moreSettings) {
+      if (serverAddressController.text.isNotEmpty) {
+        await ref.read(serverSettingsEntryProvider.notifier).changeValue(serverAddressController.text);
+      }
+      if (calendarNameController.text.isNotEmpty) {
+        await ref.read(calendarNameSettingsEntryProvider.notifier).changeValue(calendarNameController.text);
+      }
+      await ref.read(additionalParametersSettingsEntryProvider.notifier).changeValue(additionalParametersController.text);
+    }
 
     User user = User(username: usernameController.text.trim(), password: passwordController.text.trim());
-    if (await userRepository.isTestUser(user)) {
-      user = TestUser(user);
-    }
+    Calendar? calendar = await ref.read(userCalendarProvider(user).future);
+    int result = (await calendar?.get()) ?? HttpStatus.networkConnectTimeoutError;
 
-    RequestResultState loginResultState = await user.login(settingsModel.calendarUrl);
-
-    if (loginResultState != RequestResultState.success) {
-      if (context.mounted) {
-        Navigator.pop(context);
-        setState(() => this.loginResultState = loginResultState);
+    if (result == HttpStatus.ok) {
+      await ref.read(userProvider.notifier).updateUser(user);
+      if (widget.synchronizeAfterLogin) {
+        await downloadLessons(ref);
       }
-      return;
-    }
-
-    await userRepository.updateUser(user);
-    if (context.mounted) {
-      Navigator.pop(context);
-      Navigator.pop(context, true);
-    }
-
-    if (widget.synchronizeAfterLogin) {
-      ref.read(lessonRepositoryProvider).downloadLessons(
-            calendarUrl: settingsModel.calendarUrl,
-            user: user,
-          );
+      if (mounted) {
+        Navigator.pop(context, true);
+      }
+    } else {
+      setState(() {
+        loginHttpResponseCode = result;
+        waiting = false;
+      });
     }
   }
 }
