@@ -4,17 +4,17 @@ import android.appwidget.AppWidgetManager
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
-import fr.skyost.timetable.widget.TodayWidgetReceiver
+import fr.skyost.timetable.widget.TodayWidgetProvider
 
 /**
- * The BroadcastReceiver that allows to refresh the widget, lesson mode, ...
+ * The BroadcastReceiver that allows to refresh the widget, ...
  */
 class NeedUpdateReceiver : BroadcastReceiver() {
     override fun onReceive(context: Context, intent: Intent?) {
         // Refreshes the widget.
-        val updateIntent = Intent(context, TodayWidgetReceiver::class.java)
+        val updateIntent = Intent(context, TodayWidgetProvider::class.java)
         updateIntent.action = AppWidgetManager.ACTION_APPWIDGET_UPDATE
-        updateIntent.putExtra(TodayWidgetReceiver.INTENT_REFRESH_WIDGETS, true)
+        updateIntent.putExtra(TodayWidgetProvider.INTENT_REFRESH_WIDGETS, true)
         context.sendBroadcast(updateIntent)
     }
 }
