@@ -1,5 +1,6 @@
 package fr.skyost.timetable.widget
 
+import android.appwidget.AppWidgetManager
 import android.content.Intent
 import android.widget.RemoteViewsService
 
@@ -9,6 +10,7 @@ import android.widget.RemoteViewsService
 class TodayWidgetService : RemoteViewsService() {
     override fun onGetViewFactory(intent: Intent): RemoteViewsFactory {
         // We create the corresponding RemoteViews factory.
-        return TodayWidgetViewsFactory(applicationContext)
+        val widgetId = intent.getIntExtra(AppWidgetManager.EXTRA_APPWIDGET_ID, AppWidgetManager.INVALID_APPWIDGET_ID)
+        return TodayWidgetViewsFactory(applicationContext, widgetId)
     }
 }
