@@ -1,9 +1,9 @@
-import 'package:ez_localization/ez_localization.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:unicaen_timetable/i18n/translations.g.dart';
 import 'package:unicaen_timetable/intro/slides/first.dart';
 import 'package:unicaen_timetable/intro/slides/slide.dart';
+import 'package:unicaen_timetable/utils/brightness_listener.dart';
 import 'package:unicaen_timetable/widgets/slide.dart';
 
 final currentSlideProvider = ChangeNotifierProvider((ref) => ValueNotifier<Slide>(const FirstSlide()));
@@ -16,14 +16,10 @@ class IntroScaffold extends StatelessWidget {
   });
 
   @override
-  Widget build(BuildContext context) => AnnotatedRegion<SystemUiOverlayStyle>(
-        value: SystemUiOverlayStyle.light.copyWith(systemNavigationBarColor: const Color(0xFF171F29)),
-        child: Scaffold(
-          backgroundColor: const Color(0xFF2C3E50),
-          body: Padding(
-            padding: EdgeInsets.only(top: MediaQuery.of(context).padding.top),
-            child: _IntroScaffoldBody(),
-          ),
+  Widget build(BuildContext context) => Scaffold(
+        body: Padding(
+          padding: EdgeInsets.only(top: MediaQuery.of(context).padding.top),
+          child: _IntroScaffoldBody(),
         ),
       );
 }
@@ -40,78 +36,91 @@ class _IntroScaffoldBody extends ConsumerWidget {
           ref.read(currentSlideProvider).value = slide.createPreviousSlide()!;
         }
       },
-      child: Stack(
-        children: [
-          SizedBox(
-            height: MediaQuery.of(context).size.height,
-            width: MediaQuery.of(context).size.width,
-            child: Theme(
-              data: ThemeData(
-                textTheme: const TextTheme(
-                  headlineMedium: TextStyle(
-                    color: Colors.white,
-                    fontSize: 38,
-                    fontWeight: FontWeight.w100,
-                    height: 1,
-                  ),
-                  bodyMedium: TextStyle(
-                    color: Colors.white,
-                  ),
-                  bodyLarge: TextStyle(
-                    color: Colors.white,
-                    fontSize: 18,
-                  ),
-                ),
-                // textButtonTheme: TextButtonThemeData(
-                //   style: ButtonStyle(
-                //     foregroundColor: MaterialStateProperty.all<Color>(Colors.white),
-                //   ),
-                // ),
-              ),
+      child: SafeArea(
+        child: Stack(
+          children: [
+            Container(
+              padding: const EdgeInsets.only(bottom: 50),
+              height: MediaQuery.of(context).size.height,
+              width: MediaQuery.of(context).size.width,
               child: SlideWidget(
                 slide: slide,
               ),
             ),
-          ),
-          Positioned(
-            right: 0,
-            bottom: 0,
-            left: 0,
-            child: createFooter(context, ref, slide),
-          ),
-        ],
+            Positioned(
+              right: 10,
+              bottom: 0,
+              left: 18,
+              child: createFooter(context, ref, slide),
+            ),
+          ],
+        ),
       ),
     );
   }
 
   /// Creates the footer widget.
-  Widget createFooter(BuildContext context, WidgetRef ref, Slide slide) => Container(
-        color: const Color(0xFF1F2B38),
-        child: Padding(
-          padding: const EdgeInsets.only(left: 20, right: 10),
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              Text(
-                '${slide.slideIndex + 1}/${Slide.slideCount}',
-                style: TextStyle(color: Colors.white),
-              ),
-              TextButton(
-                onPressed: () async {
-                  if (slide.isLastSlide) {
-                    await Navigator.pushReplacementNamed(context, '/');
-                    // ignore: use_build_context_synchronously
-                  } else if (await slide.onGoToNextSlide(context)) {
-                    ref.read(currentSlideProvider).value = slide.createNextSlide()!;
-                  }
-                },
-                style: ButtonStyle(
-                  foregroundColor: WidgetStateProperty.all<Color>(Colors.white),
+  Widget createFooter(BuildContext context, WidgetRef ref, Slide slide) => Row(
+        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+        children: [
+          Text.rich(
+            TextSpan(
+              children: [
+                TextSpan(
+                  text: '${slide.slideIndex + 1}',
+                  style: const TextStyle(
+                    fontWeight: FontWeight.bold,
+                  ),
                 ),
-                child: Text(context.getString('intro.buttons.${slide.isLastSlide ? 'finish' : 'next'}').toUpperCase()),
-              ),
-            ],
+                const TextSpan(
+                  text: '/',
+                ),
+                const TextSpan(
+                  text: '${Slide.slideCount}',
+                  style: TextStyle(
+                    fontWeight: FontWeight.bold,
+                  ),
+                ),
+              ],
+            ),
           ),
+          _NextButton(
+            slide: slide,
+          ),
+        ],
+      );
+}
+
+/// The next button.
+class _NextButton extends ConsumerStatefulWidget {
+  /// The slide instance.
+  final Slide slide;
+
+  /// Creates a new next button instance.
+  const _NextButton({
+    required this.slide,
+  });
+
+  @override
+  ConsumerState<ConsumerStatefulWidget> createState() => _NextButtonState();
+}
+
+/// The next button state.
+class _NextButtonState extends ConsumerState<_NextButton> with BrightnessListener {
+  @override
+  Widget build(BuildContext context) => (currentBrightness == Brightness.light ? FilledButton.tonalIcon : TextButton.icon)(
+        onPressed: () async {
+          if (widget.slide.isLastSlide) {
+            await Navigator.pushReplacementNamed(context, '/');
+          } else if (await widget.slide.onGoToNextSlide(context)) {
+            ref.read(currentSlideProvider).value = widget.slide.createNextSlide()!;
+          }
+        },
+        icon: Icon(
+          widget.slide.isLastSlide ? Icons.check : Icons.chevron_right,
+        ),
+        label: Text(
+          (widget.slide.isLastSlide ? translations.intro.buttons.finish : translations.intro.buttons.next).toUpperCase(),
         ),
       );
 }
