@@ -1,42 +1,102 @@
-import 'dart:math' as math;
-
-import 'package:ez_localization/ez_localization.dart';
+import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart' hide Page;
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:flutter_widget_from_html_core/flutter_widget_from_html_core.dart';
+import 'package:unicaen_timetable/i18n/translations.g.dart';
 import 'package:unicaen_timetable/pages/page.dart';
+import 'package:unicaen_timetable/utils/utils.dart';
+import 'package:unicaen_timetable/widgets/drawer/list_title.dart';
+import 'package:unicaen_timetable/widgets/list_page.dart';
 
-/// A page that allows the user to contact me in case of any bug occurred / improvements needed.
-class BugsImprovementsPage extends Page {
-  /// The page identifier.
-  static const String id = 'bugs_improvements';
-
-  /// Creates a new bugs / improvements page instance.
-  const BugsImprovementsPage({
+/// The bugs / improvements page list tile.
+class BugsImprovementsPageListTile extends StatelessWidget {
+  /// Creates a new bugs / improvements page list tile.
+  const BugsImprovementsPageListTile({
     super.key,
-  }) : super(
-          pageId: id,
-          icon: Icons.bug_report,
-        );
+  });
 
   @override
-  Widget build(BuildContext context, WidgetRef ref) => Center(
-        child: SingleChildScrollView(
-          padding: const EdgeInsets.all(40),
-          child: Column(
-            children: [
-              Padding(
-                padding: const EdgeInsets.only(bottom: 20),
-                child: Icon(
-                  Icons.bug_report,
-                  size: math.min(100, MediaQuery.of(context).size.width),
-                  color: Colors.red[400],
-                ),
-              ),
-              HtmlWidget(context.getString('bugs_improvements.message.github')),
-              HtmlWidget(context.getString('bugs_improvements.message.website')),
-            ],
+  Widget build(BuildContext context) => PageListTitle(
+        page: BugsImprovementsPage(),
+        title: translations.bugsImprovements.title,
+        icon: const Icon(Icons.bug_report),
+      );
+}
+
+/// The bugs / improvements page app bar.
+class BugsImprovementsPageAppBar extends StatelessWidget {
+  /// Creates a new about page app bar.
+  const BugsImprovementsPageAppBar({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) => AppBar(
+        title: Text(translations.about.title),
+      );
+}
+
+/// The bugs / improvements page widget.
+class BugsImprovementsPageWidget extends StatefulWidget {
+  /// Creates a new bugs / improvements page instance.
+  const BugsImprovementsPageWidget({
+    super.key,
+  });
+
+  @override
+  State<StatefulWidget> createState() => _BugsImprovementsPageWidgetState();
+}
+
+/// The bugs / improvements page widget state.
+class _BugsImprovementsPageWidgetState extends State<BugsImprovementsPageWidget> {
+  /// The issue tracker recognizer.
+  late final TapGestureRecognizer issueTrackerRecognizer = TapGestureRecognizer()..onTap = () => Utils.openUrl('https://github.com/Skyost/UnicaenTimetable/issues/');
+
+  /// The contact form recognizer.
+  late final TapGestureRecognizer contactFormRecognizer = TapGestureRecognizer()..onTap = () => Utils.openUrl('https://skyost.eu/#contact');
+
+  @override
+  Widget build(BuildContext context) => ListPageWidget(
+        header: ListPageHeader(
+          icon: const Icon(Icons.bug_report),
+          title: Text(
+            translations.bugsImprovements.title,
           ),
         ),
+        body: ListPageBody(
+          children: [
+            Text.rich(
+              TextSpan(
+                children: [
+                  translations.bugsImprovements.message.github(
+                    issueTracker: (text) => TextSpan(
+                      text: text,
+                      recognizer: issueTrackerRecognizer,
+                      style: linkTextStyle,
+                    ),
+                  ),
+                  const TextSpan(text: ' '),
+                  translations.bugsImprovements.message.website(
+                    contactForm: (text) => TextSpan(
+                      text: text,
+                      recognizer: contactFormRecognizer,
+                      style: linkTextStyle,
+                    ),
+                  ),
+                ],
+              ),
+            ),
+          ],
+        ),
       );
+
+  /// Returns the link text style.
+  TextStyle get linkTextStyle => TextStyle(
+        decoration: TextDecoration.underline,
+        color: Theme.of(context).colorScheme.primary,
+      );
+
+  @override
+  void dispose() {
+    issueTrackerRecognizer.dispose();
+    super.dispose();
+  }
 }
