@@ -1,51 +1,184 @@
 import 'dart:io';
 
-import 'package:ez_localization/ez_localization.dart';
-import 'package:flutter/material.dart' hide Page;
+import 'package:flutter/material.dart' hide Page, DateTimeRange;
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_week_view/flutter_week_view.dart';
 import 'package:intl/intl.dart';
 import 'package:share_plus/share_plus.dart';
+import 'package:unicaen_timetable/i18n/translations.g.dart';
 import 'package:unicaen_timetable/model/lessons/lesson.dart';
 import 'package:unicaen_timetable/model/lessons/repository.dart';
-import 'package:unicaen_timetable/model/settings/entries/application/sidebar_days.dart';
-import 'package:unicaen_timetable/model/settings/settings.dart';
+import 'package:unicaen_timetable/model/settings/days_to_display.dart';
 import 'package:unicaen_timetable/pages/page.dart';
-import 'package:unicaen_timetable/pages/page_container.dart';
-import 'package:unicaen_timetable/theme.dart';
+import 'package:unicaen_timetable/utils/date_time_range.dart';
 import 'package:unicaen_timetable/utils/utils.dart';
+import 'package:unicaen_timetable/widgets/drawer/list_title.dart';
 import 'package:unicaen_timetable/widgets/flutter_week_view.dart';
 
-/// The page that allows to show a day's lessons.
-class DayViewPage extends FlutterWeekViewWidget {
+/// The day view page list tile.
+class DayViewPageListTile extends StatelessWidget {
+  /// The week day.
+  final int day;
+
+  /// Creates a new day view page list tile.
+  const DayViewPageListTile({
+    super.key,
+    required this.day,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    DateTime monday = DateTime.now().atMonday;
+    DateTime date = monday.add(Duration(days: day - 1));
+    return PageListTitle(
+      page: DayViewPage(day: day),
+      title: DateFormat.EEEE(TranslationProvider.of(context).locale.languageCode).format(date).capitalize(),
+      icon: _DayViewPageListTileIcon(
+        day: day,
+      ),
+    );
+  }
+}
+
+/// A day view page list tile icon, with a number.
+class _DayViewPageListTileIcon extends StatelessWidget {
+  /// The week day.
+  final int day;
+
+  /// Creates a new day view page list tile icon.
+  const _DayViewPageListTileIcon({
+    required this.day,
+  });
+
+  @override
+  Widget build(BuildContext context) => Stack(
+        children: [
+          const Icon(Icons.square_rounded),
+          Positioned.fill(
+            child: Center(
+              child: Text(
+                day.toString(),
+                style: TextStyle(
+                  color: Theme.of(context).scaffoldBackgroundColor,
+                ),
+                textAlign: TextAlign.center,
+              ),
+            ),
+          ),
+        ],
+      );
+}
+
+/// The about week view app bar.
+class DayViewPageAppBar extends ConsumerWidget {
   /// The share button key.
   final GlobalKey _shareButtonKey = GlobalKey();
 
   /// The week day.
-  final int weekDay;
+  final int day;
 
-  /// Creates a new day view page instance.
-  DayViewPage({
+  /// Creates a new week view page app bar.
+  DayViewPageAppBar({
     super.key,
-    required this.weekDay,
-    required super.pageId,
-    super.icon,
+    required this.day,
   });
 
-  /// Builds a day view page id.
-  static String buildPageId(int weekDay) => 'day_view_$weekDay';
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    List<int> sidebarDays = ref.watch(daysToDisplayEntryProvider).valueOrNull ?? [];
+    DateTime monday = DateTime.now().atMonday;
+    DateTime date = monday.add(Duration(days: day - 1));
+    return AppBar(
+      title: Text(DateFormat.EEEE(TranslationProvider.of(context).locale.languageCode).format(date).capitalize()),
+      actions: [
+        if (sidebarDays.isNotEmpty)
+          IconButton(
+            icon: Icon(Platform.isAndroid ? Icons.arrow_back : Icons.arrow_back_ios),
+            onPressed: () => DayViewPageWidget._previousDay(ref, day),
+          ),
+        if (sidebarDays.isNotEmpty)
+          IconButton(
+            icon: Icon(Platform.isAndroid ? Icons.arrow_forward : Icons.arrow_forward_ios),
+            onPressed: () => DayViewPageWidget._nextDay(ref, day),
+          ),
+        const WeekPickerButton(),
+        IconButton(
+          key: _shareButtonKey,
+          icon: const Icon(Icons.share),
+          onPressed: () async {
+            String? languageCode = TranslationProvider.of(context).locale.languageCode;
+            RenderObject? renderObject = _shareButtonKey.currentContext?.findRenderObject();
+            Offset? position;
+
+            if (renderObject is RenderBox) {
+              position = renderObject.localToGlobal(Offset.zero);
+            }
+
+            StringBuffer builder = StringBuffer();
+            DateTime monday = ref.watch(dateProvider);
+            DateTime day = monday.add(Duration(days: this.day));
+            List<Lesson> lessons = await ref.read(lessonsProvider(DateTimeRange.oneDay(day)).future);
+            builder.write('${DateFormat.yMd(languageCode).format(date)} :\n\n');
+            for (Lesson lesson in lessons) {
+              builder.write('$lesson\n');
+            }
+            String content = builder.toString();
+            await Share.share(
+              content.substring(0, content.lastIndexOf('\n')),
+              sharePositionOrigin: position == null ? null : Rect.fromLTWH(position.dx, position.dy, 24, 40),
+            );
+          },
+        ),
+      ],
+    );
+  }
+}
+
+/// The day view page widget.
+class DayViewPageWidget extends ConsumerStatefulWidget {
+  /// The week day.
+  final int day;
+
+  /// Creates a new day view page instance.
+  const DayViewPageWidget({
+    super.key,
+    required this.day,
+  });
 
   @override
-  String buildTitle(BuildContext context) {
-    DateTime monday = DateTime.now().atMonday;
-    DateTime date = monday.add(Duration(days: weekDay - 1));
-    return DateFormat.EEEE(EzLocalization.of(context)?.locale.languageCode).format(date).capitalize();
+  ConsumerState<ConsumerStatefulWidget> createState() => _DayViewPageWidgetState();
+
+  /// Goes to the previous day.
+  static Future<void> _previousDay(WidgetRef ref, int currentWeekDay) async {
+    List<int> sidebarDays = await ref.read(daysToDisplayEntryProvider.future);
+    int previousDay = sidebarDays.previousDay(currentWeekDay);
+    if (previousDay >= currentWeekDay) {
+      DateTime monday = ref.read(dateProvider);
+      monday = monday.subtract(const Duration(days: 7));
+      ref.read(dateProvider.notifier).changeDate(monday);
+    }
+    ref.read(pageProvider.notifier).changePage(DayViewPage(day: previousDay));
   }
 
+  /// Goes to the next day.
+  static Future<void> _nextDay(WidgetRef ref, int currentWeekDay) async {
+    List<int> sidebarDays = await ref.read(daysToDisplayEntryProvider.future);
+    int nextDay = sidebarDays.nextDay(currentWeekDay);
+    if (nextDay <= currentWeekDay) {
+      DateTime monday = ref.read(dateProvider);
+      monday = monday.add(const Duration(days: 7));
+      ref.read(dateProvider.notifier).changeDate(monday);
+    }
+    ref.read(pageProvider.notifier).changePage(DayViewPage(day: nextDay));
+  }
+}
+
+/// The day view page widget state.
+class _DayViewPageWidgetState extends FlutterWeekViewWidgetState<DayViewPageWidget> {
   @override
-  Widget buildChild(BuildContext context, WidgetRef ref, List<FlutterWeekViewEvent> events) {
-    DateTime date = resolveDate(ref);
-    UnicaenTimetableTheme theme = ref.watch(settingsModelProvider).resolveTheme(context);
+  Widget buildChild(List<FlutterWeekViewEventWithLesson> events) {
+    DateTime monday = ref.watch(dateProvider);
+    DateTime date = monday.add(Duration(days: widget.day - 1));
     return GestureDetector(
       onHorizontalDragEnd: (details) {
         if (details.primaryVelocity == null) {
@@ -53,211 +186,29 @@ class DayViewPage extends FlutterWeekViewWidget {
         }
 
         if (details.primaryVelocity! > 0) {
-          previousDay(ref);
+          DayViewPageWidget._previousDay(ref, widget.day);
         }
 
         if (details.primaryVelocity! < 0) {
-          nextDay(ref);
+          DayViewPageWidget._nextDay(ref, widget.day);
         }
       },
-      child: DayView(
+      child: DayView<FlutterWeekViewEventWithLesson>(
         date: date,
         events: events,
-        initialTime: const HourMinute(hour: 7),
-        style: theme.createDayViewStyle(date),
-        dayBarStyle: theme.createDayBarStyle(date, (year, month, day) => formatDate(context, year, month, day)),
-        hoursColumnStyle: theme.createHoursColumnStyle(),
+        initialTime: const TimeOfDay(hour: 7, minute: 0),
+        style: createDayViewStyle(date),
+        dayBarStyle: createDayBarStyle(date, (year, month, day) => formatDate(context, year, month, day)),
+        hourColumnStyle: createHoursColumnStyle(),
+        eventWidgetBuilder: createEventWidget,
       ),
     );
   }
 
   @override
-  Future<List<FlutterWeekViewEvent>> createEvents(BuildContext context, WidgetRef ref) async {
-    SettingsModel settingsModel = ref.watch(settingsModelProvider);
-    LessonRepository lessonRepository = ref.watch(lessonRepositoryProvider);
-    DateTime date = resolveDate(ref);
-    return (await lessonRepository.getLessonsForDate(date)).map((lesson) => createEvent(context, lesson, lessonRepository, settingsModel)).toList();
+  AsyncValue<List<LessonWithColor>> queryLessons() {
+    DateTime monday = ref.watch(dateProvider);
+    DateTime day = monday.add(Duration(days: widget.day - 1));
+    return ref.watch(lessonsProvider(DateTimeRange.oneDay(day)));
   }
-
-  /// Goes to the previous day.
-  void previousDay(WidgetRef ref) {
-    SidebarDaysSettingsEntry sidebarDays = ref.read(settingsModelProvider).sidebarDaysEntry;
-    int previousDay = sidebarDays.previousDay(weekDay);
-    if (previousDay >= weekDay) {
-      ValueNotifier<DateTime> monday = ref.read(currentDateProvider);
-      monday.value = monday.value.subtract(const Duration(days: 7));
-    }
-    ref.read(currentPageProvider).value = buildPageId(previousDay);
-  }
-
-  /// Goes to the next day.
-  void nextDay(WidgetRef ref) {
-    SidebarDaysSettingsEntry sidebarDays = ref.read(settingsModelProvider).sidebarDaysEntry;
-    int nextDay = sidebarDays.nextDay(weekDay);
-    if (nextDay <= weekDay) {
-      ValueNotifier<DateTime> monday = ref.read(currentDateProvider);
-      monday.value = monday.value.add(const Duration(days: 7));
-    }
-    ref.read(currentPageProvider).value = buildPageId(nextDay);
-  }
-
-  /// Resolves the date from the given ref.
-  DateTime resolveDate(WidgetRef ref, {bool listen = true}) {
-    DateTime monday = (listen ? ref.watch(currentDateProvider) : ref.read(currentDateProvider)).value;
-    return monday.add(Duration(days: weekDay - 1));
-  }
-
-  @override
-  List<Widget> buildActions(BuildContext context, WidgetRef ref) {
-    SidebarDaysSettingsEntry sidebarDays = ref.read(settingsModelProvider).sidebarDaysEntry;
-    return [
-      if (sidebarDays.value.isNotEmpty)
-        IconButton(
-          icon: Icon(Platform.isAndroid ? Icons.arrow_back : Icons.arrow_back_ios),
-          onPressed: () => previousDay(ref),
-        ),
-      if (sidebarDays.value.isNotEmpty)
-        IconButton(
-          icon: Icon(Platform.isAndroid ? Icons.arrow_forward : Icons.arrow_forward_ios),
-          onPressed: () => nextDay(ref),
-        ),
-      const WeekPickerButton(),
-      IconButton(
-        key: _shareButtonKey,
-        icon: const Icon(Icons.share),
-        onPressed: () async {
-          String? languageCode = EzLocalization.of(context)?.locale.languageCode;
-          RenderObject? renderObject = _shareButtonKey.currentContext?.findRenderObject();
-          Offset? position;
-
-          if (renderObject is RenderBox) {
-            position = renderObject.localToGlobal(Offset.zero);
-          }
-
-          StringBuffer builder = StringBuffer();
-          DateTime date = resolveDate(ref, listen: false);
-          LessonRepository lessonRepository = ref.read(lessonRepositoryProvider);
-          List<Lesson> lessons = await lessonRepository.getLessonsForDate(date)
-            ..sort();
-          builder.write('${DateFormat.yMd(languageCode).format(date)} :\n\n');
-          for (Lesson lesson in lessons) {
-            builder.write('$lesson\n');
-          }
-          String content = builder.toString();
-          await Share.share(
-            content.substring(0, content.lastIndexOf('\n')),
-            sharePositionOrigin: position == null ? null : Rect.fromLTWH(position.dx, position.dy, 24, 40),
-          );
-        },
-      ),
-    ];
-  }
-
-  @override
-  bool isSamePage(Page other) => super.isSamePage(other) && other is DayViewPage && weekDay == other.weekDay;
-}
-
-/// Monday day view page.
-class MondayPage extends DayViewPage {
-  /// The page identifier.
-  static const String id = 'day_view_1';
-
-  /// Creates a new monday page instance.
-  MondayPage({
-    super.key,
-  }) : super(
-          weekDay: DateTime.monday,
-          pageId: id,
-          icon: const IconData(0xf03a4, fontFamily: 'MaterialCommunityIcons'),
-        );
-}
-
-/// Tuesday day view page.
-class TuesdayPage extends DayViewPage {
-  /// The page identifier.
-  static const String id = 'day_view_2';
-
-  /// Creates a new tuesday page instance.
-  TuesdayPage({
-    super.key,
-  }) : super(
-          weekDay: DateTime.tuesday,
-          pageId: id,
-          icon: const IconData(0xf03a7, fontFamily: 'MaterialCommunityIcons'),
-        );
-}
-
-/// Wednesday day view page.
-class WednesdayPage extends DayViewPage {
-  /// The page identifier.
-  static const String id = 'day_view_3';
-
-  /// Creates a new wednesday page instance.
-  WednesdayPage({
-    super.key,
-  }) : super(
-          weekDay: DateTime.wednesday,
-          pageId: id,
-          icon: const IconData(0xf03aa, fontFamily: 'MaterialCommunityIcons'),
-        );
-}
-
-/// Thursday day view page.
-class ThursdayPage extends DayViewPage {
-  /// The page identifier.
-  static const String id = 'day_view_4';
-
-  /// Creates a new thursday page instance.
-  ThursdayPage({
-    super.key,
-  }) : super(
-          weekDay: DateTime.thursday,
-          pageId: id,
-          icon: const IconData(0xf03ad, fontFamily: 'MaterialCommunityIcons'),
-        );
-}
-
-/// Friday day view page.
-class FridayPage extends DayViewPage {
-  /// The page identifier.
-  static const String id = 'day_view_5';
-
-  /// Creates a new friday page instance.
-  FridayPage({
-    super.key,
-  }) : super(
-          weekDay: DateTime.friday,
-          pageId: id,
-          icon: const IconData(0xf03b1, fontFamily: 'MaterialCommunityIcons'),
-        );
-}
-
-/// Saturday day view page.
-class SaturdayPage extends DayViewPage {
-  /// The page identifier.
-  static const String id = 'day_view_6';
-
-  /// Creates a new saturday page instance.
-  SaturdayPage({
-    super.key,
-  }) : super(
-          weekDay: DateTime.saturday,
-          pageId: id,
-          icon: const IconData(0xf03b3, fontFamily: 'MaterialCommunityIcons'),
-        );
-}
-
-/// Sunday day view page.
-class SundayPage extends DayViewPage {
-  /// The page identifier.
-  static const String id = 'day_view_7';
-
-  /// Creates a new sunday page instance.
-  SundayPage({
-    super.key,
-  }) : super(
-          weekDay: DateTime.sunday,
-          pageId: id,
-          icon: const IconData(0xf03b6, fontFamily: 'MaterialCommunityIcons'),
-        );
 }
